trying to understand my mistake

+++ lib/calculas.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-double df(double (*func)(double),double x){
+double df(double (*func)(int),double x){
     // double eps = 1.61011979e-4;
     double eps = 1.61012e-4;
     double df = (func(x+eps) -func(x-eps))/(2*eps);
